requirements.txt
Flask==0.11.1
Pillow==3.3.1
